EV-17 create DraftEventCard, add new logic of rendering functionality for EventCard
